requirements.txt
discord.py>=2.3.2
quart>=0.19.4
python-dotenv>=1.0.0
aiohttp>=3.9.1
